requirements.txt
Django
django-crispy-forms
pillow
django-ckeditor